ai_hedge_fund/requirements.txt
# AI Hedge Fund System Dependencies

# Core framework
pyautogen>=0.2.0
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Data handling and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Financial data and analysis
alpaca-trade-api>=3.0.0
alpaca-py>=0.20.0
polygon-api-client>=1.10.0
pandas-datareader>=0.10.0
TA-Lib>=0.4.25
quantlib>=1.31

# Database and storage
duckdb>=0.9.0
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0
psycopg2-binary>=2.9.0

# API and web
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
requests>=2.31.0
aiohttp>=3.8.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0
grafana-api>=1.0.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
httpx>=0.24.0  # for testing FastAPI

# Development tools
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter and analysis
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Async and concurrency
asyncio-mqtt>=0.13.0
celery>=5.3.0

# Cryptography and security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Date and time handling
python-dateutil>=2.8.0
pytz>=2023.3

# Mathematical and statistical
statsmodels>=0.14.0
arch>=6.2.0  # For GARCH models

# Machine Learning (optional advanced features)
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0

# Backtesting
backtrader>=1.9.78.123
zipline-reloaded>=2.2.0

# Risk management
pyfolio>=0.9.2
empyrical>=0.5.5

# Alternative data sources
tweepy>=4.14.0  # Twitter API
feedparser>=6.0.10  # RSS feeds
beautifulsoup4>=4.12.0  # Web scraping

# Docker and deployment
gunicorn>=21.2.0
docker>=6.1.0

# Monitoring and observability
sentry-sdk>=1.29.0
datadog>=0.47.0

# Performance optimization
cython>=3.0.0
numba>=0.57.0

# Excel and reporting
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Networking and messaging
websockets>=11.0.0
pika>=1.3.0  # RabbitMQ

# Utility libraries
tqdm>=4.65.0
joblib>=1.3.0
python-slugify>=8.0.0
marshmallow>=3.20.0